Major refactor
Since we've decided to keep the two projects separated it's necessary
that if some one wants to clone the repository he has a clear idea of
what's going on here: in src/lib folder there are the three main
components of the library libgrive (drive, protocol and util); while
in the src/cli folder there's the command line interface tool that uses
the library. The CLI is called in the same way as the library (grive).

In this way, we can keep multiple clients using the same library in just
one repo..